backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
paho-mqtt>=2.1.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.1
apscheduler>=3.10.4
sqlalchemy>=2.0.36
aiosqlite>=0.20.0